# Parquet/Arrow: implement faster spatial filtering with ArrowArray interface