--- a/ogr/ogrsf_frmts/ogrsf_frmts.h
+++ b/ogr/ogrsf_frmts/ogrsf_frmts.h
@@ -158,6 +158,10 @@ class CPL_DLL OGRLayer : public GDALMajorObject
     CreateSchemaForWKBGeometryColumn(const OGRGeomFieldDefn *poFieldDefn,
                                      const char *pszArrowFormat = "z");
 
+    bool CanPostFilterArrowArray(const struct ArrowSchema *schema) const;
+    void PostFilterArrowArray(const struct ArrowSchema *schema,
+                              struct ArrowArray *array) const;
+
   public:
     OGRLayer();
     virtual ~OGRLayer();