--- a/swig/include/gdal_array.i
+++ b/swig/include/gdal_array.i
@@ -1514,7 +1514,7 @@ PyObject* _RecordBatchAsNumpy(VoidPtrAsLong recordBatchPtr,
                 Py_RETURN_NONE;
             }
             const int32_t* offsetsToBytes = (const int32_t*)arrayField->children[0]->buffers[1] + static_cast<size_t>(arrayField->children[0]->offset);
-            const char* bytes = (const char*)arrayField->children[0]->buffers[2] + static_cast<size_t>(arrayField->children[0]->offset);
+            const char* bytes = (const char*)arrayField->children[0]->buffers[2];
             numpyArray = PyArray_SimpleNew(1, &dims, NPY_OBJECT);
             for( npy_intp j = 0; j < dims; j++ )
             {
@@ -1596,7 +1596,7 @@ PyObject* _RecordBatchAsNumpy(VoidPtrAsLong recordBatchPtr,
                 Py_RETURN_NONE;
             }
             const int32_t* offsetsToBytes = (const int32_t*)arrayField->children[0]->buffers[1] + static_cast<size_t>(arrayField->children[0]->offset);
-            const char* bytes = (const char*)arrayField->children[0]->buffers[2] + static_cast<size_t>(arrayField->children[0]->offset);
+            const char* bytes = (const char*)arrayField->children[0]->buffers[2];
             numpyArray = PyArray_SimpleNew(1, &dims, NPY_OBJECT);
             for( npy_intp j = 0; j < dims; j++ )
             {