… autotest/ogr/data/parquet/wkt_with_dict.parquet
… ogr/ogrsf_frmts/generic/ogrlayer.cpp
… ogr/ogrsf_frmts/generic/ogrlayerarrow.cpp