We no longer fallback to the slow & generic implementation that goes
through GetNextFeature(), but directly post filter the ArrowArray to
remove features not intersecting the spatial filter.

The performance gain is mostly when a big number of features is selected
(the fallback GetNextFeature() has already an efficient spatial
filtering, so when selecting a small number of features, this
optimization doesn't bring anything)
Can be up to 10x faster in that situation.

Now:
```

$ time bench_ogr_batch nz-building-outlines.parquet -spat 1167513 4794680 2089113 6190596

real    0m1,275s
user    0m1,565s
sys     0m0,322s
```

Before:
```
$ time bench_ogr_batch nz-building-outlines.parquet -spat 1167513 4794680 2089113 6190596

real    0m13,507s
user    0m13,728s
sys     0m0,712s

```